Rename module to openopc2
