+++ openopc2/exceptions.py
