+++ openopc2/tests/test_list.py
