+++ openopc2/tests/test_opc_com.py
@@ -2,7 +2,7 @@ import time
 from unittest import TestCase, skipIf
 
 from opc_server_config import OPC_HOST, OPC_SERVER, USE_GATEWAY, OPC_CLASS
-from openopc.opc_da_com import OpcCom
+from openopc2.opc_da_com import OpcCom
 
 TAG = 'Bucket Brigade.Int1'
