+++ openopc2/tests/test_properties.py
